+++ summon-core/src/commonMain/kotlin/code/yousef/summon/modifier/TransitionModifiers.kt
@@ -62,6 +62,8 @@ fun Modifier.transitionDelay(value: Number): Modifier =
 fun Modifier.hover(hoverModifier: Modifier): Modifier =
     hover(hoverModifier.styles)
 
+@Suppress("EXTENSION_SHADOWED_BY_MEMBER")
+@Deprecated("Extension shadowed by member function", level = DeprecationLevel.HIDDEN)
 fun Modifier.pointerEvents(value: String): Modifier =
     style("pointer-events", value)